apply suggestions and rename import-fe folder
