+++ src/i18n/en.json
@@ -433,7 +433,7 @@
       "value": "Value",
       "importedFrom": {
         "baseEmpreinte": "Base Empreinte logo",
-        "negaoctet": "NegaOctet logo",
+        "negaOctet": "NegaOctet logo",
         "manual": "Manual"
       },
       "post": "Post:",
