+++ src/scripts/negaOctet/getEmissionFactors.ts
@@ -1,11 +1,11 @@
 import { Command } from 'commander'
-import { getEmissionFactorsFromCSV } from '../../services/import-fe/negaoctet/getEmissionFactors'
+import { getEmissionFactorsFromCSV } from '../../services/importEmissionFactor/negaOctet/getEmissionFactors'
 
 const program = new Command()
 
 program
-  .name('base-empreinte-import-emission')
-  .description("Script pour importer les facteurs d'émission depuis la base negaoctet")
+  .name('negaoctet-import-emission')
+  .description("Script pour importer les facteurs d'émission depuis la base NegaOctet")
   .version('1.0.0')
   .requiredOption('-n, --name <value>', 'Nom de la version')
   .requiredOption('-f, --file <value>', 'Import from CSV file')
